add i2c addresses to headers

+++ scd4x.c
@@ -77,7 +77,7 @@ int scd4x_get_serial(scd4x_t *scd4x, char* serial) {
 }
 
 int scd4x_get_result_float(scd4x_t *scd4x, uint16_t *co2_ppm, float *temp_C,
-		uint16_t *humidity_percent) {
+		float *humidity_percent) {
 	int status;
 	scd4x_result_t result = {0};
 	uint16_t cmd;
