+++ scd4x.h
@@ -58,7 +58,7 @@ typedef struct {
 #define SCD4X_CMD_WAKE_UP 						htobe16(0x36F6)
 
 int scd4x_get_result_float(scd4x_t *scd4x, uint16_t *co2_ppm, float *temp_C,
-		uint16_t *humidity_percent);
+		float *humidity_percent);
 int scd4x_start(scd4x_t *scd4x);
 
 #endif /* SCD4X_H_ */
